add `static_cast<int>(…)` for downcast warnings

+++ ogles_gpgpu/common/proc/fifo.cpp
@@ -112,7 +112,7 @@ void FifoProc::process(int position, Logger logger) {
 }
 
 ProcInterface* FifoProc::operator[](int i) const {
-    int index = modulo(m_outputIndex + i, procPasses.size());
+    int index = modulo(static_cast<int>(m_outputIndex) + i, static_cast<int>(procPasses.size()));
     return procPasses[index];
 }
 
@@ -192,13 +192,13 @@ void FifoProc::createFBOTex(bool genMipmap) {
 int FifoProc::render(int position) {
     // Render into input FBO
     if (m_count == int(size())) {
-        m_outputIndex = modulo(m_inputIndex + 1, size());
+        m_outputIndex = modulo(static_cast<int>(m_inputIndex) + 1, static_cast<int>(size()));
     }
 
     getInputFilter()->render();
 
     m_count = std::min(m_count + 1, int(size()));
-    m_inputIndex = modulo(m_inputIndex + 1, size());
+    m_inputIndex = modulo(static_cast<int>(m_inputIndex) + 1, static_cast<int>(size()));
 
     return 0;
 }
